backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
PyMySQL
Flask-Cors
python-dotenv
